Decouple and reorganize headers
